% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe1.R
\name{pe1}
\alias{pe1}
\title{The Parseval series estimation of pi  using R}
\usage{
pe1(n)
}
\arguments{
\item{n}{the number of series  terms}
}
\value{
A compound number
}
\description{
The series estimation of pi using Parseval law
}
\examples{
\dontrun{
R1<-p_e1(100)
}
}
