% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe2.R
\name{pe2}
\alias{pe2}
\title{The series estimation of pi  using R}
\usage{
pe2(n)
}
\arguments{
\item{n}{the number of series  terms}
}
\value{
A compound number
}
\description{
The Leibniz series estimation of pi using Parseval law
}
\examples{
\dontrun{
R1<-p_e2(100)
}
}
