\name{SC19097-package}
\alias{SC19097-package}
\alias{SC19097}
\docType{package}
\title{
  Two estimation methods of pis
}
\description{
  Two estimation methods of pi (Leibniz and Parseval).
}
\details{
  We can get the approximate value of the first N-term series of different methods.
}
\author{
shenmaoquan 
Maintainer:shenmaoquan <329567885@qq.com>
}
\keyword{ package }

